modified chatbot settings panel background
