+++ packages/gen-ai/frontend/src/app/Chatbot/components/ChatbotSettingsPanel.tsx
@@ -20,6 +20,7 @@ import { ACCORDION_ITEMS } from '~/app/Chatbot/const';
 import useAccordionState from '~/app/Chatbot/hooks/useAccordionState';
 import { UseSourceManagementReturn } from '~/app/Chatbot/hooks/useSourceManagement';
 import { UseFileManagementReturn } from '~/app/Chatbot/hooks/useFileManagement';
+import useDarkMode from '~/app/Chatbot/hooks/useDarkMode';
 import { useMCPSelectionContext } from '~/app/context/MCPSelectionContext';
 import MCPServersPanelWithContext from '~/app/Chatbot/mcp/MCPServersPanelWithContext';
 import UploadedFilesList from './UploadedFilesList';
@@ -60,6 +61,7 @@ const ChatbotSettingsPanel: React.FunctionComponent<ChatbotSettingsPanelProps> =
 }) => {
   const accordionState = useAccordionState();
   const { selectedServersCount, saveSelectedServersToPlayground } = useMCPSelectionContext();
+  const isDarkMode = useDarkMode();
 
   const SETTINGS_PANEL_WIDTH = 'chatbot-settings-panel-width';
   const DEFAULT_WIDTH = '460px';
@@ -98,12 +100,26 @@ const ChatbotSettingsPanel: React.FunctionComponent<ChatbotSettingsPanelProps> =
             <AccordionToggle
               onClick={() => accordionState.onAccordionToggle(ACCORDION_ITEMS.MODEL_DETAILS)}
               id={ACCORDION_ITEMS.MODEL_DETAILS}
+              style={{
+                backgroundColor: isDarkMode
+                  ? 'var(--pf-v6-c-page__main-section--BackgroundColor)'
+                  : 'var(--pf-t--global--background--color--100)',
+              }}
             >
               <Title headingLevel="h2" size="lg">
                 Model details
               </Title>
             </AccordionToggle>
-            <AccordionContent id="model-details-content">
+            <AccordionContent
+              id="model-details-content"
+              style={{
+                backgroundColor: isDarkMode
+                  ? 'var(--pf-v6-c-page__main-section--BackgroundColor)'
+                  : 'var(--pf-t--global--background--color--100)',
+                margin: '0',
+                padding: '0.5rem',
+              }}
+            >
               <Form>
                 <FormGroup label="Model" fieldId="model-details">
                   <ModelDetailsDropdown
@@ -145,11 +161,18 @@ const ChatbotSettingsPanel: React.FunctionComponent<ChatbotSettingsPanelProps> =
             <AccordionToggle
               onClick={() => accordionState.onAccordionToggle(ACCORDION_ITEMS.SOURCES)}
               id={ACCORDION_ITEMS.SOURCES}
+              style={{
+                backgroundColor: isDarkMode
+                  ? 'var(--pf-v6-c-page__main-section--BackgroundColor)'
+                  : 'var(--pf-t--global--background--color--100)',
+              }}
             >
               <Flex
                 justifyContent={{ default: 'justifyContentSpaceBetween' }}
                 alignItems={{ default: 'alignItemsCenter' }}
-                style={{ width: '100%' }}
+                style={{
+                  width: '100%',
+                }}
               >
                 <FlexItem>
                   <Title headingLevel="h2" size="lg">
@@ -183,7 +206,16 @@ const ChatbotSettingsPanel: React.FunctionComponent<ChatbotSettingsPanelProps> =
                 </FlexItem>
               </Flex>
             </AccordionToggle>
-            <AccordionContent id="sources-content">
+            <AccordionContent
+              id="sources-content"
+              style={{
+                backgroundColor: isDarkMode
+                  ? 'var(--pf-v6-c-page__main-section--BackgroundColor)'
+                  : 'var(--pf-t--global--background--color--100)',
+                margin: '0',
+                padding: '0.5rem',
+              }}
+            >
               <Form>
                 <FormGroup fieldId="sources">
                   <ChatbotSourceUploadPanel
@@ -218,6 +250,11 @@ const ChatbotSettingsPanel: React.FunctionComponent<ChatbotSettingsPanelProps> =
             <AccordionToggle
               onClick={() => accordionState.onAccordionToggle(ACCORDION_ITEMS.MCP_SERVERS)}
               id={ACCORDION_ITEMS.MCP_SERVERS}
+              style={{
+                backgroundColor: isDarkMode
+                  ? 'var(--pf-v6-c-page__main-section--BackgroundColor)'
+                  : 'var(--pf-t--global--background--color--100)',
+              }}
             >
               <Flex alignItems={{ default: 'alignItemsCenter' }}>
                 <FlexItem>
@@ -227,14 +264,22 @@ const ChatbotSettingsPanel: React.FunctionComponent<ChatbotSettingsPanelProps> =
                 </FlexItem>
                 {selectedServersCount > 0 && (
                   <FlexItem>
-                    <Label key={1} color="blue" className="pf-v6-u-ml-sm">
+                    <Label key={1} color="blue">
                       {selectedServersCount}
                     </Label>
                   </FlexItem>
                 )}
               </Flex>
             </AccordionToggle>
-            <AccordionContent id="mcp-servers-content">
+            <AccordionContent
+              id="mcp-servers-content"
+              style={{
+                backgroundColor: isDarkMode
+                  ? 'var(--pf-v6-c-page__main-section--BackgroundColor)'
+                  : 'var(--pf-t--global--background--color--100)',
+                margin: '0',
+              }}
+            >
               <MCPServersPanelWithContext onSelectionChange={saveSelectedServersToPlayground} />
             </AccordionContent>
           </AccordionItem>
